requirements.txt
streamlit==1.31.0
openai==1.12.0
pinecone-client==3.0.0
PyPDF2==3.0.1
pdfplumber==0.10.3
pdf2image==1.17.0
pytesseract==0.3.10
Pillow==10.2.0
python-dotenv==1.0.0
